Refactoring: - Changes related to Schema change (new column address instead of street and buildingNumber) - create packages for User and Order entities - in case of Gender was not found return null
