+++ src/main/java/com/photostudio/dao/jdbc/mapper/OrderRowMapper.java
@@ -1,6 +1,6 @@
 package com.photostudio.dao.jdbc.mapper;
 
-import com.photostudio.entity.Order;
+import com.photostudio.entity.order.Order;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
