+++ src/main/java/com/photostudio/entity/user/UserRole.java
@@ -1,4 +1,4 @@
-package com.photostudio.entity;
+package com.photostudio.entity.user;
 
 public enum UserRole {
     USER("USER"),
