+++ src/main/java/com/photostudio/service/OrderService.java
@@ -1,6 +1,6 @@
 package com.photostudio.service;
 
-import com.photostudio.entity.Order;
+import com.photostudio.entity.order.Order;
 
 import java.util.List;
